Rename PrimoButton to PalaButton
- Rename PrimoButton.svelte to PalaButton.svelte
- Update imports and component references
- Maintain consistency with Pala naming
